resolved circular dependencies between router and main lib based on skott graph
